+++ src/builtin_entity_parser.rs
@@ -60,7 +60,6 @@ impl BuiltinEntityParserLoader {
             Some(parser_path) => Some(GazetteerParser::from_path(parser_path)?),
             None => None,
         };
-
         Ok(BuiltinEntityParser {
             gazetteer_parser,
             rustling_parser,
@@ -99,7 +98,6 @@ impl BuiltinEntityParser {
             })
             .flat_map(|kind| kind.try_ontology_into().ok())
             .collect::<Vec<OutputKind>>();
-
         let rustling_entities = if rustling_output_kinds.is_empty() {
             vec![]
         } else {
@@ -277,13 +275,63 @@ mod test {
     fn test_entities_extraction() {
         let parser = BuiltinEntityParserLoader::new(Language::EN).load().unwrap();
         assert_eq!(
-            vec![BuiltinEntityKind::Number, BuiltinEntityKind::Time],
+            vec![BuiltinEntityKind::Number, BuiltinEntityKind::Date],
             parser
-                .extract_entities("Book me restaurant for two people tomorrow", None)
+                .extract_entities("Book me a restaurant for two people tomorrow", None)
                 .unwrap()
                 .iter()
                 .map(|e| e.entity_kind)
                 .collect_vec()
+        );
+
+        assert_eq!(
+            vec![BuiltinEntityKind::Datetime],
+            parser
+                .extract_entities("Book me a restaurant for tomorrow", Some(&[BuiltinEntityKind::Datetime]))
+                .unwrap()
+                .iter()
+                .map(|e| e.entity_kind)
+                .collect_vec()
+        );
+
+        assert_eq!(
+            vec![BuiltinEntityKind::Datetime],
+            parser
+                .extract_entities("Book me a restaurant for tomorrow at 8pm", None)
+                .unwrap()
+                .iter()
+                .map(|e| e.entity_kind)
+                .collect_vec()
+        );
+
+        assert_eq!(
+            vec![BuiltinEntityKind::Date],
+            parser
+                .extract_entities("Book me a restaurant for tomorrow", Some(&[BuiltinEntityKind::Date]))
+                .unwrap()
+                .iter()
+                .map(|e| e.entity_kind)
+                .collect_vec()
+        );
+
+        assert_eq!(
+            vec![BuiltinEntityKind::TimePeriod],
+            parser
+                .extract_entities("Book the meeting room from 10am to 11am", None)
+                .unwrap()
+                .iter()
+                .map(|e| e.entity_kind)
+                .collect_vec()
+        );
+
+        assert_eq!(
+            vec![BuiltinEntityKind::Time, BuiltinEntityKind::Time],
+            parser
+                .extract_entities("Book the meeting room from 10am to 11am", Some(&[BuiltinEntityKind::Time]))
+            .unwrap()
+            .iter()
+            .map(|e| e.entity_kind)
+            .collect_vec()
         );
 
         assert_eq!(
@@ -416,7 +464,7 @@ mod test {
     #[test]
     fn test_entities_extraction_for_non_space_separated_languages() {
         let parser = BuiltinEntityParserLoader::new(Language::JA).load().unwrap();
-        let expected_time_value = InstantTimeValue {
+        let expected_datetime_value = InstantTimeValue {
             value: "2013-02-10 00:00:00 +01:00".to_string(),
             grain: Grain::Day,
             precision: Precision::Exact,
@@ -425,8 +473,8 @@ mod test {
         let expected_entity = BuiltinEntity {
             value: "二 千 十三 年二 月十 日".to_string(),
             range: 10..24,
-            entity_kind: BuiltinEntityKind::Time,
-            entity: InstantTime(expected_time_value.clone()),
+            entity_kind: BuiltinEntityKind::Datetime,
+            entity: InstantTime(expected_datetime_value.clone()),
         };
 
         let parsed_entities = parser.extract_entities(
@@ -439,9 +487,9 @@ mod test {
         assert_eq!(expected_entity.range, parsed_entity.range);
         assert_eq!(expected_entity.entity_kind, parsed_entity.entity_kind);
 
-        if let SlotValue::InstantTime(ref parsed_time) = parsed_entity.entity {
-            assert_eq!(expected_time_value.grain, parsed_time.grain);
-            assert_eq!(expected_time_value.precision, parsed_time.precision);
+        if let SlotValue::InstantTime(ref parsed_datetime) = parsed_entity.entity {
+            assert_eq!(expected_datetime_value.grain, parsed_datetime.grain);
+            assert_eq!(expected_datetime_value.precision, parsed_datetime.precision);
         } else {
             panic!("")
         }